requirements.txt
selenium
webdriver-manager
requests
google-generativeai
beautifulsoup4
apscheduler
python-dotenv
pandas
psutil